Fixing a bug with byte swapping of the port.
The DNSService routines require the port to be in network byte
order.  On a PowerPC, this is the same as the host byte order, but
when executed on an Intel platform, network byte order and host
byte order differ.

Both the register and resolve routines have been modified to use
htons() and ntohs() to byte swap the port number.