Added a new "register_with_name" tcl command
I ran into a case where I need to specify the service name, rather
than let the dns_sd library choose it for me.  Since I didn't want
to break any users already using the ::register command, I added
a new command called ::register_with_name whose first argument is
the service name.  All of the following arguments are the same.

Having said that, what we should probably do is make some sort of
switch on the register command so that we can have one command
but not affect backwards compatibility.  Something like:

 ::bonjour::register -name "My service" _foo._tcp 9000 {key value}

that way, the existing syntax would still be acceptable and the
"-name" argument would be optional.  I thought about doing this
but couldn't think of a good way to do it without restructuring
the C code a bit, so I figured I would leave it to the package
maintainer.

I also updated the server.tcl example script to show the usage of
the new command.