Adding the "-name" option to the register command.
An issue was reported where a user needed to specify a service name rather
than using the default provided by the dns_sd library.  The new "-name"
option, when passed to the register command, allows this value to be specified.

The server.tcl and client.tcl scripts have also been updated to demonstrate
usage of this new option.