Adding handling of errors in DNSService* functions.
The library will now return an error when a call to one of the
DNSService* functions fails.  The error returned is a list consisting
of an error message, the name of the DNSService* function generating
the error, and the error code (usually a string).

Whenever a background error occurs in a DNSService callback, the error
message will match the format for other functions.

Closes #5.